+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Deprecated: `"emptyLineBefore"` option for `declaration-block-properties-order`. If you use this option, please consider creating a plugin for the community.
 - Deprecated: `"hierarchicalSelectors"` option for `indentation`.  If you use this option, please consider creating a plugin for the community.
 - Fixed: the string formatter no longer errors on non-rule errors.
+- Fixed: `selector-type-no-unknown` now ignores obsolete HTML tags and `<hgroup>`.
 - Fixed: `selector-max-compound-selectors` no longer errors on Less mixins.
 - Fixed: `selector-list-comma-*` rules now ignore Less mixins.
 - Fixed: `selector-type-no-unknown` now ignores all *An+B notation* and linguistic pseudo-classes.